rename `protocol::serialization` to `boxed_format`
