+++ manul/src/protocol.rs
@@ -11,13 +11,14 @@ to be executed by a [`Session`](`crate::session::Session`).
 For more details, see the documentation of the mentioned traits.
 */
 
+mod boxed_format;
 mod boxed_round;
 mod errors;
 mod message;
 mod round;
 mod round_id;
-mod serialization;
 
+pub use boxed_format::BoxedFormat;
 pub use boxed_round::BoxedRound;
 pub use errors::{
     DeserializationError, DirectMessageError, EchoBroadcastError, LocalError, MessageValidationError,
@@ -29,7 +30,6 @@ pub use round::{
     Payload, Protocol, ProtocolError, RequiredMessageParts, RequiredMessages, Round,
 };
 pub use round_id::{RoundId, TransitionInfo};
-pub use serialization::BoxedFormat;
 
 pub(crate) use errors::ReceiveErrorType;
 pub(crate) use message::ProtocolMessagePartHashable;
