+++ manul/src/protocol/boxed_format.rs
